requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.11.1
gunicorn==19.7.1
implicit==0.2.5
packaging==16.8
psycopg2==2.7.1
py-postgresql==1.2.1
pyparsing==2.2.0
pytz==2017.2
scipy==0.19.0
six==1.10.0
virtualenv==15.1.0
whitenoise==3.3.0
django-rest==0.0.1
djangorestframework==3.5.3

